Update README.md
docs: change to README

Added three new images for the data which better 
display what data was used in this project. The first
image is the wave plot, the second image is a 
matplotlib spectrogram, and the third image is
a custom created stft rgb image. All three images
are made using the same audio data.

Included information about where the three minute
versions of the audio files are.